Combine container/component code for same components
