--- app/src/containers/AppContainer.jsx
@@ -1,15 +1,0 @@
-import { connect } from 'react-redux'
-import { checkLogin } from '../actions/auth'
-import { App } from '../components'
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    checkLogin: () => dispatch(checkLogin())
-  }
-}
-
-const AppContainer = connect(
-  null, // no mapStateToProps
-mapDispatchToProps)(App)
-
-export default AppContainer
