+++ app/src/containers/index.js
@@ -1,4 +1,4 @@
-export AppContainer from './AppContainer.jsx'
-export NavBarContainer from './NavBarContainer.jsx'
-export LandingContainer from './LandingContainer.jsx'
+export App from './App.jsx'
+export NavBar from './NavBar.jsx'
+export LandingPage from './LandingPage.jsx'
 export HomePage from './HomePage.jsx'
