+++ app/src/routes.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Route, IndexRoute } from 'react-router'
-import { AppContainer, LandingContainer, HomePage } from './containers'
-import { AboutPage, NotFoundPage, Landing, Itinerary } from './components'
+import { App, LandingPage, HomePage } from './containers'
+import { AboutPage, NotFoundPage, Itinerary } from './components'
 
 import AuthService from './utils/AuthService'
 
@@ -14,8 +14,8 @@ const requireAuth = (nextState, replace) => {
 
 export default function createRoutes() {
   return(
-    <Route path='/' component={AppContainer}>
-      <IndexRoute component={LandingContainer} />
+    <Route path='/' component={App}>
+      <IndexRoute component={LandingPage} />
       <Route path='/home' component={HomePage} onEnter={requireAuth}/>
       <Route path="/itinerary" component={Itinerary} onEnter={requireAuth} />
       <Route path='*' component={NotFoundPage} />
